requirements.txt
discord.py==2.7.1
httpx==0.28.1
python-dotenv==1.1.1
